Relocate Remote Registration Service
