+++ vcell-client/src/main/java/org/vcell/util/network/RemoteRegistrationService.java
@@ -1,8 +1,9 @@
-package org.vcell.service.registration.remote;
+package org.vcell.util.network;
 
 import cbit.vcell.message.server.bootstrap.client.RemoteProxyVCellConnectionFactory.RemoteProxyException;
 import cbit.vcell.resource.PropertyLoader;
 import org.vcell.api.client.VCellApiClient;
+import org.vcell.api.utils.DTOOldAPI;
 import org.vcell.service.registration.RegistrationService;
 import org.vcell.util.DataAccessException;
 import org.vcell.util.UseridIDExistsException;
@@ -22,8 +23,8 @@ public class RemoteRegistrationService implements RegistrationService {
 	@Override
 	public UserInfo insertUserInfo(UserInfo newUserInfo, boolean bUpdate) throws RemoteProxyException, DataAccessException, UseridIDExistsException {
 		try (VCellApiClient apiClient = getvCellApiClient()){
-			org.vcell.api.common.UserInfo apiUserInfo = apiClient.insertUserInfo(newUserInfo.getApiUserInfo());
-			return UserInfo.fromApiUserInfo(apiUserInfo);
+			org.vcell.api.common.UserInfo apiUserInfo = apiClient.insertUserInfo(DTOOldAPI.getApiUserInfo(newUserInfo));
+			return DTOOldAPI.fromApiUserInfo(apiUserInfo);
 		} catch (IOException | NoSuchAlgorithmException | KeyStoreException | KeyManagementException e) {
 			throw new RemoteProxyException("failed to insert user: "+e.getMessage(), e);
 		}
